Restructure functions and methods; enforce separation of concerns
